requirements.txt
streamlit
pandas
folium
streamlit-folium
numpy
plotly
gTTS
google-generativeai